requirements.txt
backoff==2.2.1
black==23.10.0
blinker==1.6.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.0
Flask-Cors==4.0.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
highlight-io==0.6.5
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mypy-extensions==1.0.0
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-logging==0.41b0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
protobuf==4.25.1
pyodbc==5.0.1
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.0
wrapt==1.16.0
zipp==3.17.0
